# Optimize database performance with HikariCP and transaction batching