--- a/.github/runner-files/ci-gradle.properties
+++ b/.github/runner-files/ci-gradle.properties
@@ -1,7 +1,9 @@
 org.gradle.daemon=false
-org.gradle.jvmargs=-Xmx5120m
+org.gradle.jvmargs=-Xmx5120m -XX:MaxMetaspaceSize=1024m
 org.gradle.workers.max=5
 org.gradle.parallel=true
+org.gradle.caching=true
 
 kotlin.incremental=false
 kotlin.compiler.execution.strategy=in-process
+kotlin.daemon.jvmargs=-Xmx3072m