--- a/server/src/main/kotlin/suwayomi/tachidesk/manga/impl/util/DirName.kt
+++ b/server/src/main/kotlin/suwayomi/tachidesk/manga/impl/util/DirName.kt
@@ -7,7 +7,6 @@ package suwayomi.tachidesk.manga.impl.util
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import org.jetbrains.exposed.sql.ResultRow
 import org.jetbrains.exposed.sql.selectAll
 import org.jetbrains.exposed.sql.transactions.transaction
 import suwayomi.tachidesk.manga.impl.util.source.GetCatalogueSource
@@ -20,31 +19,38 @@ import java.io.File
 
 private val applicationDirs: ApplicationDirs by injectLazy()
 
-private fun getMangaDir(mangaId: Int): String {
-    val mangaEntry = getMangaEntry(mangaId)
-    val source = GetCatalogueSource.getCatalogueSourceOrStub(mangaEntry[MangaTable.sourceReference])
+private fun getMangaDir(mangaId: Int): String =
+    transaction {
+        val mangaEntry = MangaTable.selectAll().where { MangaTable.id eq mangaId }.first()
+        val source = GetCatalogueSource.getCatalogueSourceOrStub(mangaEntry[MangaTable.sourceReference])
 
-    val sourceDir = SafePath.buildValidFilename(source.toString())
-    val mangaDir = SafePath.buildValidFilename(mangaEntry[MangaTable.title])
-    return "$sourceDir/$mangaDir"
-}
+        val sourceDir = SafePath.buildValidFilename(source.toString())
+        val mangaDir = SafePath.buildValidFilename(mangaEntry[MangaTable.title])
+        "$sourceDir/$mangaDir"
+    }
 
 private fun getChapterDir(
     mangaId: Int,
     chapterId: Int,
-): String {
-    val chapterEntry = transaction { ChapterTable.selectAll().where { ChapterTable.id eq chapterId }.first() }
+): String =
+    transaction {
+        // Get chapter data and build chapter-specific directory name
+        val chapterEntry = ChapterTable.selectAll().where { ChapterTable.id eq chapterId }.first()
 
-    val chapterDir =
-        SafePath.buildValidFilename(
-            when {
-                chapterEntry[ChapterTable.scanlator] != null -> "${chapterEntry[ChapterTable.scanlator]}_${chapterEntry[ChapterTable.name]}"
-                else -> chapterEntry[ChapterTable.name]
-            },
-        )
+        val chapterDir =
+            SafePath.buildValidFilename(
+                when {
+                    chapterEntry[ChapterTable.scanlator] != null -> {
+                        "${chapterEntry[ChapterTable.scanlator]}_${chapterEntry[ChapterTable.name]}"
+                    }
+                    else -> chapterEntry[ChapterTable.name]
+                },
+            )
 
-    return getMangaDir(mangaId) + "/$chapterDir"
-}
+        // Get manga directory and combine with chapter directory
+        // Note: This creates a nested transaction, but Exposed handles this with useNestedTransactions=true
+        getMangaDir(mangaId) + "/$chapterDir"
+    }
 
 fun getThumbnailDownloadPath(mangaId: Int): String = applicationDirs.thumbnailDownloadsRoot + "/$mangaId"
 
@@ -70,16 +76,21 @@ fun updateMangaDownloadDir(
     mangaId: Int,
     newTitle: String,
 ): Boolean {
-    val mangaEntry = getMangaEntry(mangaId)
-    val source = GetCatalogueSource.getCatalogueSourceOrStub(mangaEntry[MangaTable.sourceReference])
+    // Get current manga directory (uses its own transaction)
+    val currentMangaDir = getMangaDir(mangaId)
 
-    val sourceDir = SafePath.buildValidFilename(source.toString())
-    val mangaDir = SafePath.buildValidFilename(mangaEntry[MangaTable.title])
+    // Build new directory path
+    val newMangaDir =
+        transaction {
+            val mangaEntry = MangaTable.selectAll().where { MangaTable.id eq mangaId }.first()
+            val source = GetCatalogueSource.getCatalogueSourceOrStub(mangaEntry[MangaTable.sourceReference])
+            val sourceDir = SafePath.buildValidFilename(source.toString())
+            val newMangaDirName = SafePath.buildValidFilename(newTitle)
+            "$sourceDir/$newMangaDirName"
+        }
 
-    val newMangaDir = SafePath.buildValidFilename(newTitle)
-
-    val oldDir = "${applicationDirs.downloadsRoot}/$sourceDir/$mangaDir"
-    val newDir = "${applicationDirs.downloadsRoot}/$sourceDir/$newMangaDir"
+    val oldDir = "${applicationDirs.downloadsRoot}/$currentMangaDir"
+    val newDir = "${applicationDirs.downloadsRoot}/$newMangaDir"
 
     val oldDirFile = File(oldDir)
     val newDirFile = File(newDir)
@@ -90,5 +101,3 @@ fun updateMangaDownloadDir(
         true
     }
 }
-
-private fun getMangaEntry(mangaId: Int): ResultRow = transaction { MangaTable.selectAll().where { MangaTable.id eq mangaId }.first() }