## Summary
Running Suwayomi-Server on Raspberry Pi 5 with a moderately large database, chapter operations were taking up to a minute to complete, making the system essentially unusable for basic reading. After profiling, I identified severe database connection and transaction bottlenecks in core functions.

This PR addresses the highest priority performance issues to make basic reading operations responsive. There are likely more optimization opportunities throughout the codebase, but these changes target the most critical bottlenecks affecting everyday usage.

## Changes
- **Database Connection Pooling**: Add HikariCP-7.0.2 with Raspberry Pi optimized settings (6 max connections, 2 min idle)
- **Transaction Batching**: Consolidate separate database transactions in DirName, ChapterForDownload, and ChapterDownloadHelper
- **Query Optimization**: Remove duplicate queries and create shared query functions
- **Code Cleanup**: Remove 6 unused methods from ChapterForDownload class
- **Build Optimization**: Configure memory settings for compilation

## Performance Results

### Method-Level Timing Comparison

**DirName.kt Functions:**
| Function | Before | After | Improvement |
|----------|--------|-------|-------------|
| getMangaDir() | 29,190ms (29.2s) | 970ms (0.97s) | **96.7% faster** |
| getChapterDir() | 27,750ms (27.8s) | 1,510ms (1.51s) | **94.6% faster** |
| getThumbnailDownloadPath() | 56,940ms (56.9s) | 10ms (0.01s) | **99.98% faster** |
| getChapterDownloadPath() | 21,090ms (21.1s) | 1,520ms (1.52s) | **92.8% faster** |

**ChapterDownloadHelper.kt Functions:**
| Function | Before | After | Improvement |
|----------|--------|-------|-------------|
| getImage() | 54,830ms (54.8s) | 300ms (0.30s) | **99.5% faster** |
| getImageCount() | 2,140ms (2.14s) | 1,310ms (1.31s) | **38.8% faster** |
| provider() | 26,090ms (26.1s) | 950ms (0.95s) | **96.4% faster** |

**ChapterForDownload.kt Functions:**
| Function | Before | After | Improvement |
|----------|--------|-------|-------------|
| asDownloadReady() | 2,140ms (2.14s) | 1,360ms (1.36s) | **36.4% faster** |
| Transaction operations | 790-1,180ms | 30-40ms | **96-97% faster** |
| Database queries | 390ms | 20ms | **94.9% faster** |

### System-Wide Improvements
- **Overall execution time**: 242s → 60s (**75% faster**)
- **Database activity samples**: 35,285 → 4,127 (**88.3% reduction**)
- **Thread blocking time**: 3.44s avg → 25.1ms avg (**99% faster**)
- **Thread parking time**: 1.29s avg → 230ms avg (**82% faster**)

The system transforms from completely unusable (60+ second load times) to highly responsive with sub-second chapter loading times.