@renjfk 2025-09-21
Here are the JFRs that I've captured [JFRs.zip](https://github.com/user-attachments/files/22454937/JFRs.zip)