@Caio99BR 2025-09-22
This is a change worth priority and bump version, no?